refactor: change to calculateRoundAPR to be reused on other components
